More modified

+++ src/components/pages/appontment/Appointment.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
-const Appointment = ({data}) => {
+const Appointment = ({data, setappointment}) => {
 
 
 
@@ -9,11 +10,42 @@ const Appointment = ({data}) => {
 
     const handleDelete = (_id) => {
 
-        fetch (`http://localhost:4000/appointments/${_id}`, {
-            method :  'DELETE',
-        })
-        .then (rse => rse.json())
-        .then (data => console.log  (data))
+
+
+
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+
+
+
+                fetch (`http://localhost:4000/appointments/${_id}`, {
+                    method :  'DELETE',
+                })
+                .then (rse => rse.json())
+                .then (data =>{ setappointment  (data)
+                
+                
+              Swal.fire(
+                'Deleted!',
+                'Your file has been deleted.',
+                'success'
+              )})
+
+
+
+
+            }
+          })
+
+
